refactor: simplify StatsGrid component by using Redux for mainStats and remove local state management
